convert wasmprinter and wasm-encoder

+++ crates/wasmprinter/src/operator.rs
@@ -9,7 +9,7 @@ use wasmparser::{
 };
 
 pub struct OperatorState {
-    op_offset: usize,
+    op_offset: u64,
     nesting_start: u32,
     label: u32,
     label_indices: Vec<u32>,
@@ -38,7 +38,7 @@ struct FoldedInstruction {
     plain: String,
     folded: Vec<FoldedInstruction>,
     results: u32,
-    offset: usize,
+    offset: u64,
 }
 
 struct Block {
@@ -47,8 +47,8 @@ struct Block {
     plain: String,
     folded: Vec<FoldedInstruction>,
     predicate: Option<Vec<FoldedInstruction>>,
-    consequent: Option<(Vec<FoldedInstruction>, usize)>,
-    offset: usize,
+    consequent: Option<(Vec<FoldedInstruction>, u64)>,
+    offset: u64,
 }
 
 pub struct PrintOperatorFolded<'printer, 'state, 'a, 'b> {
@@ -1430,8 +1430,8 @@ impl<'a> VisitSimdOperator<'a> for PrintOperator<'_, '_, '_, '_> {
 }
 
 pub trait OpPrinter {
-    fn branch_hint(&mut self, offset: usize, taken: bool) -> Result<()>;
-    fn set_offset(&mut self, offset: usize);
+    fn branch_hint(&mut self, offset: u64, taken: bool) -> Result<()>;
+    fn set_offset(&mut self, offset: u64);
     fn visit_operator(
         &mut self,
         reader: &mut OperatorsReader<'_>,
@@ -1442,7 +1442,7 @@ pub trait OpPrinter {
 }
 
 impl OpPrinter for PrintOperator<'_, '_, '_, '_> {
-    fn branch_hint(&mut self, offset: usize, taken: bool) -> Result<()> {
+    fn branch_hint(&mut self, offset: u64, taken: bool) -> Result<()> {
         self.printer.newline(offset)?;
         let desc = if taken { "\"\\01\"" } else { "\"\\00\"" };
         self.printer.result.start_comment()?;
@@ -1451,7 +1451,7 @@ impl OpPrinter for PrintOperator<'_, '_, '_, '_> {
         Ok(())
     }
 
-    fn set_offset(&mut self, offset: usize) {
+    fn set_offset(&mut self, offset: u64) {
         self.operator_state.op_offset = offset;
     }
 
@@ -1486,7 +1486,7 @@ impl OpPrinter for PrintOperator<'_, '_, '_, '_> {
 }
 
 impl OpPrinter for PrintOperatorFolded<'_, '_, '_, '_> {
-    fn branch_hint(&mut self, offset: usize, taken: bool) -> Result<()> {
+    fn branch_hint(&mut self, offset: u64, taken: bool) -> Result<()> {
         let mut hint = String::new();
         hint.push_str("@metadata.code.branch_hint ");
         hint.push_str(if taken { "\"\\01\"" } else { "\"\\00\"" });
@@ -1499,7 +1499,7 @@ impl OpPrinter for PrintOperatorFolded<'_, '_, '_, '_> {
         Ok(())
     }
 
-    fn set_offset(&mut self, offset: usize) {
+    fn set_offset(&mut self, offset: u64) {
         self.operator_state.op_offset = offset;
     }
